pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Ultimate FastAPI Tutorial"
authors = ["ChristopherGS"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.18.3"
fastapi = "~0.68.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "^3.0.1"
SQLAlchemy = "^1.4.22"
alembic = "^1.6.5"
tenacity = "^8.0.1"
greenlet = "^1.1.2"
PyMySQL = "^1.0.2"
cryptography = "^37.0.4"
asyncio = "^3.4.3"
python-dotenv = "^0.20.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.3"
boto3 = "^1.24.51"
gunicorn = "^20.1.0"
web3 = "^5.30.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"